exercise/core/views.py
from rest_framework import viewsets
from .models import Recipe
from .serializers import RecipeSerializer


class RecipeViewSet(viewsets.ModelViewSet):
    """Manage Recipes in the database"""

    queryset = Recipe.objects.all()
    serializer_class = RecipeSerializer

    def get_queryset(self):
        queryset = self.queryset
        name = self.request.query_params.get("name")

        if name is not None:
            queryset = queryset.filter(name__contains=name)

        return queryset
